chore: (Android) Remove remnants of v1 results (#4514)
#### Details

To facilitate with migration, the feature to add ATFA rules shipped with a feature flag that allowed v1 or v2 results to be selected. Now that the feature has shipped, and before starting work on changing the service to a non-HTTP protocol, the feature flag and the ability to produce/consume v1 results needs to be removed.

##### Motivation

Preparation for the protocol change feature.

##### Context

<!-- Are there any parts that you've intentionally left out-of-scope for a later PR to handle? -->
The commits took a "peel the onion" approach from the outside in. It might be simpler to review on a per-commit basis. Note that the `ColorContrast` rule was kept around for V1 compatibility, even though it was excluded from V2 results. It goes away completely in this PR

<!-- Were there any alternative approaches you considered? What tradeoffs did you consider? -->

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a" in the checkbox -->
- [ ] Addresses an existing issue: #0000
- [ ] Ran `yarn fastpass`
- [ ] Added/updated relevant unit test(s) (and ran `yarn test`)
- [ ] Verified code coverage for the changes made. Check coverage report at: `<rootDir>/test-results/unit/coverage`
- [ ] PR title *AND* final merge commit title both start with a semantic tag (`fix:`, `chore:`, `feat(feature-name):`, `refactor:`). See `CONTRIBUTING.md`.
- [ ] (UI changes only) Added screenshots/GIFs to description above
- [ ] (UI changes only) Verified usability with NVDA/JAWS