refactor: accessibility-insights-ui's package.json (#4515)
Details
Similar to the refactor from #3519, this PR moves the contents from /src/packages/accessibility-insights-ui/root to /packages/ui and updates relevant package.json files. Also updates /packages/README.md to reflect current package layout.

When running yarn build:package:ui locally from this branch, /packages/ui/drop/ is created and contains index.js and ui.css

Motivation
The change to workingDir in pipeline/build-package.template.yaml caused the npm publish step to stop working for the accessibility-insights-ui package, because it could no longer find the package.json. With this refactor, package.json will be in the specified directory, /packages/ui.

Context
Considered adding a "style" field to package.json for the ui.css file, similar to the typings field added in #3519 for the report package, but it appears to not be standardized and could cause breaks in the future