feat(download-report-json): Only show to JSON export option in assessment (#4767)
#### Details
This PR follows up on #4754 to only show the new to JSON export option in assessment, not fastpass. To accomplish this, `ExportDropdown` was refactored to expect a list of currently applicable `ReportExportServices` rather than a `ReportExportServiceProvider`. `ReportExportServiceProvider` was updated to provide lists of `ReportExportServices` for FastPass and Assessment. Props/deps were updated such that `ReportExportServiceProvider` is now passed down only until the point at which `ExportDialog` diverges for FastPass and Assessment (specifically in `ReportExportDialogFactory` for web and `ReportExportComponent` for unified). At this point, it is used to generate the appropriate list of `ReportExportServices`, which is then passed through until it is used in `ExportDialog` and `ExportDropdown`.

##### Motivation
To JSON is only an export option for assessment.

##### Context
- The to JSON option still doesn't do anything
- Some slight refactoring was done in `ExportDialog` to ensure that a dropdown with only one option isn't displayed in fastpass if the export JSON feature flag is enabled and export to codepen feature flag is not

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a" in the checkbox -->
- [n/a] Addresses an existing issue: #0000
- [x] Ran `yarn fastpass`
- [x] Added/updated relevant unit test(s) (and ran `yarn test`)
- [x] Verified code coverage for the changes made. Check coverage report at: `<rootDir>/test-results/unit/coverage`
- [x] PR title *AND* final merge commit title both start with a semantic tag (`fix:`, `chore:`, `feat(feature-name):`, `refactor:`). See `CONTRIBUTING.md`.
- [n/a] (UI changes only) Added screenshots/GIFs to description above
- [n/a] (UI changes only) Verified usability with NVDA/JAWS