feat(download-report-json): Modify export telemetry to include export service (#4774)
#### Details
This PR updates the telemetry logged on report export click in two ways:
1. `exportResultsService` is added to the event with potential values of `html`, `json`, and `codepen`
2. `exportResultsData` is removed from the event (read more in Context)

##### Motivation
Telemetry updates for the json export feature

##### Context
In theory, `exportResultsData` was recording the length of the downloaded html string. In reality, it had a bug which caused it to record the length of the _previously downloaded_ html string from that session, which was usually 0. This PR fixes that bug by removing the value entirely (per discussion with @ferBonnin, we don't need that data).

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a" in the checkbox -->
- [n/a] Addresses an existing issue: #0000
- [x] Ran `yarn fastpass`
- [x] Added/updated relevant unit test(s) (and ran `yarn test`)
- [x] Verified code coverage for the changes made. Check coverage report at: `<rootDir>/test-results/unit/coverage`
- [x] PR title *AND* final merge commit title both start with a semantic tag (`fix:`, `chore:`, `feat(feature-name):`, `refactor:`). See `CONTRIBUTING.md`.
- [n/a] (UI changes only) Added screenshots/GIFs to description above
- [n/a] (UI changes only) Verified usability with NVDA/JAWS