fix: Use aria-label on NavLinkButton to improve Orca reading behavior (#4904)
#### Details
Currently, we set aria-current="page" on the left nav bar button for the currently selected page. This is right and proper. Unfortunately, the Orca screen reader ignores this setting in both the Web and Android products. To improve the user's experience, this PR replaces the use of aria-current with an aria-label to simulate the proper reading behavior. 

##### Motivation
Addresses accessibility bug in Linux.

##### Context
I considered leaving the aria-current value as-is, but that results in a double reading of current page from non-Orca screen readers.

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a" in the checkbox -->
- [x] Addresses an existing issue: ADO 1886192
- [x] Ran `yarn fastpass`
- [x] Added/updated relevant unit test(s) (and ran `yarn test`)
- [x] Verified code coverage for the changes made. Check coverage report at: `<rootDir>/test-results/unit/coverage`
- [x] PR title *AND* final merge commit title both start with a semantic tag (`fix:`, `chore:`, `feat(feature-name):`, `refactor:`). See `CONTRIBUTING.md`.
- [n/a] (UI changes only) Added screenshots/GIFs to description above
- [x] (UI changes only) Verified usability with NVDA/JAWS