fix: link-in-text-block accessibility issue (#4932)
Details
Change text colors from secondary-text to primary-text to meet the minimum contrast ratio of 3:1 when links are found in text blocks. Using primary-text provides a contrast ratio of 3.345:1 between the surrounding text and link text colors.

Motivation
Fixes #4928

Context
Changed the text color for descriptions of GenericToggle. This additionally affects the description text in "Enable high contrast" under the settings panel, as well as any descriptions for the toggles under preview features. I decided to go with this, rather than selectively changing the text in the areas with accessibility issues, so the UI looks more consistent (same text color for toggle descriptions throughout) and prevents similar accessibility issues in the future.