feat(tabstops-report): Update TabStopsFailedInstanceSection to render properly in report (#5035)
#### Details
This PR brings the behavior of `TabStopsFailedInstanceSection` closer to what we desire in `FastPassReport`. At a high level, there are two key changes:
- Instances (comments) are now rendered in the report. If there are no instances for a failure, a message is displayed
- The section will now always be displayed, even if there are no failures

Previously, instances were not rendering because the `DetailsList` component does not render properly in our report. This PR introduces the `TabStopsInstanceSectionPropsFactory`, which allows `TabStopsRequirementsWithInstances` to behave differently in the Report than in DetailsView. Existing logic pertaining to instance rendering was moved out of `TabStopsRequirementsWithInstances` into the `TabStopsInstanceSectionPropsFactory` for FastPass, which is then passed in through deps. The `TabStopsInstanceSectionPropsFactory` for Report contains new configuration and logic which leverages `AssessmentReportInstanceList` to render instances.

This PR should have no impact on the `TabStopsFailedInstanceSection` in the extension itself.

Screenshot of the updated report:
![Screenshot of the updated FastPass report](https://user-images.githubusercontent.com/4615491/148621041-2980aea7-ca0f-4d1f-81bf-8785d3463f63.png)

##### Motivation
Continuing to bring the `FastPassReport` closer to spec for the feature

##### Context
I initially started creating an parallel version of the `TabStopsFailedInstanceSection` for the report rather than modifying the existing one to allow for reuse. This resulted a lot of almost-duplicate components that would have to be kept in sync down the road, so I instead opted to refactor for reuse.

I thought about the (more straightforward) approach of passing in a different component for collapsible content. However, there were a few more differences between the report and extension `TabStopsFailedInstanceSection`s (such as how to render failures with no instances and button clicking behavior or lack thereof), so that wouldn't work.

Out of scope is updating styling of the tab stops section of the report to match the figma, including fixing its ordering so failures are displayed first.

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a" in the checkbox -->
- [n/a] Addresses an existing issue: #0000
- [x] Ran `yarn fastpass`
- [x] Added/updated relevant unit test(s) (and ran `yarn test`)
- [x] Verified code coverage for the changes made. Check coverage report at: `<rootDir>/test-results/unit/coverage`
- [x] PR title *AND* final merge commit title both start with a semantic tag (`fix:`, `chore:`, `feat(feature-name):`, `refactor:`). See `CONTRIBUTING.md`.
- [x] (UI changes only) Added screenshots/GIFs to description above
- [x] (UI changes only) Verified usability with NVDA/JAWS