feat(tabstops-report) Update failed counter on fastpass report (#5052)
#### Details
This PR makes two updates to the FastPassReport:
1. It updates the failed instance count for tab stops to count each failed requirement with no user comments as 1 towards the requirement and overall section count
2. It italicizes the "This requirement has failed but no comment has been added" message for failed requirements with no instances

To accomplish 1, `TabStopsFailedCounter` has been split into two implementations: one for the extension details view that maintains existing behavior and one for the report that counts no instance requirements as described above. This PR has no impact on the behavior of the extension details view.

Screenshot showing both changes:
![Screenshot showing both changes](https://user-images.githubusercontent.com/4615491/149244174-1ced6204-b037-47ca-89f4-9dff79810906.png)


##### Motivation
Bringing report closer to spec

##### Context
I considered breaking the no comment message out into its own component, but it seemed to trivial to deserve that. Happy to do so if anyone feels differently.

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a" in the checkbox -->
- [n/a] Addresses an existing issue: #0000
- [x] Ran `yarn fastpass`
- [x] Added/updated relevant unit test(s) (and ran `yarn test`)
- [x] Verified code coverage for the changes made. Check coverage report at: `<rootDir>/test-results/unit/coverage`
- [x] PR title *AND* final merge commit title both start with a semantic tag (`fix:`, `chore:`, `feat(feature-name):`, `refactor:`). See `CONTRIBUTING.md`.
- [x] (UI changes only) Added screenshots/GIFs to description above
- [n/a] (UI changes only) Verified usability with NVDA/JAWS