fix(tabstops-report): fix requirement/comment table widths and spacing (#5115)
#### Details

This PR addresses a fit and finish issues in the new `Details view > FastPass > Tab stops` UI. Before this PR, at page widths of 600px-704px, one or both of the new `DetailsList`-based tables on the page would end up using undesirable horizontal scroll bars. Also, the icons for comments "edit" and "delete" controls weren't quite lined up as intended with the "Pass/Fail" choice group in the preceding table. This PR fixes the spacing issues such that the icons line up and horizontal scroll bars are avoided.

Note that, due to an implementation detail of the fabric `DetailsView`, there are still horizontal scroll bars that can appear as one is in the process of resizing the window; they disappear as soon as the user stops draggin the resize handle. This is a consequence of Fabric using scripted calculations of width rather than CSS width properties; it's not something we can realistically fix short of avoiding their `DetailsList` control entirely.

#### Screenshots

Taken at 600px width (the lowest width before we collapse the left nav bar, ie, the lowest width the tables ever render at).

Before:

![screenshot showing both tables with horizontal scrollbars and comment buttons not lined up](https://user-images.githubusercontent.com/376284/151270985-12c18cd0-4e36-4746-b367-581098792d81.png)

After:

![screenshot showing both tables with no horizontal scrollbars and comment buttons lined up](https://user-images.githubusercontent.com/376284/151270714-9d9d889c-ccb0-488a-90bc-409d86e25537.png)

##### Motivation

See #5099

##### Context

Note that the updated spacing still slightly mismatches the Figma in that the Figma suggests a wider gap between the word "Comment" and the user's actual comment; I left that as-is with the thinner spacing because there isn't enough room to flow the user-specified text otherwise. @ferBonnin and I discussed that we might want to update this case to wrap more similarly to how the requirement descriptions do in the future (ie, using a structure like `<div><span>Comment:</span> The user's comment</div>` in one block element instead of the current version with `<div>Comment</div><div>The user's comment</div>` that has them wrapping separately). However, if we do that, we'd want to update the similar cases in the Assessment simultaneously for consistency, so leaving that out of scope for this PR.

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a" in the checkbox -->
- [x] Addresses an existing issue: #5099 (partial)
- [x] Ran `yarn fastpass`
- [x] Added/updated relevant unit test(s) (and ran `yarn test`)
- [x] Verified code coverage for the changes made. Check coverage report at: `<rootDir>/test-results/unit/coverage`
- [x] PR title *AND* final merge commit title both start with a semantic tag (`fix:`, `chore:`, `feat(feature-name):`, `refactor:`). See `CONTRIBUTING.md`.
- [x] (UI changes only) Added screenshots/GIFs to description above
- [x] (UI changes only) Verified usability with NVDA/JAWS