fix(tabstops-auto-target-page-vis): visual helper toggle bug adding failure instances (#5164)
Details
Fixes the bug described in #5113, in which toggling the visual helper off and then back on again adds duplicate failure instances to the list in the details view.

Motivation
Addresses bug and part of feature work.

Context
Uses the needToCollectTabbingResults flag to check if it is the first time results have been added. Does not add tab stop instances if needToCollectTabbingResults is false (gets set to false in onResultsChangedState in analyzer-controller.ts).