feat(tabstops-auto-target-page-vis): Add TabStopsRequirementResultProcessor (#5219)
Details
Create TabStopsRequirementResultProcessor to take some of the responsibility of the result processing away from the TabStopsAnalyzer and AnalyzerController.

Motivation
Part of feature work.

Context
AnalyzerController included a function that was used only in the TabStops case and was calling teardown() when it should not have. TabStopsAnalyzer also had too much responsibility in processing results and managing the requirement runner. This PR moves these functionalities to a separate class, TabStopsRequirementResultProcessor, to fix these issues. This also fixes an issue that #5202 tried to address, in which the keyboard navigation errors weren't being added to the store/failure instances list.