chore(deps): update moment to 2.29.2 (#5339)
#### Details

Resolves component governance noise related to CVE-2022-24785. The issue doesn't affect our usage, but there's no reason not to update to resolve the alert.

The important update here is `moment` to 2.29.2. It was previously pinned to specific version 2.29.1 via dependency chain `appium-adb > @appium/support > moment`, so the update was implemented by removing and re-adding the `appium-adb` package, which updated `@appium/support` to a version which pins to the desired `moment` version.

The `webpack.config` update is to handle a case where a new transitive dependency (`read-pkg`) introduced by the new `@appium/support` version happens to invoke `require('some_file.json')`. This is valid, but caused webpack errors because we configured webpack not to allow `json` resolution. The config update allows webpack to handle this case.

#### Motivation
Resolve alert noise

#### Context
n/a

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a" in the checkbox -->
- [n/a] Addresses an existing issue: #0000
- [n/a] Ran `yarn fastpass`
- [n/a] Added/updated relevant unit test(s) (and ran `yarn test`)
- [n/a] Verified code coverage for the changes made. Check coverage report at: `<rootDir>/test-results/unit/coverage`
- [x] PR title *AND* final merge commit title both start with a semantic tag (`fix:`, `chore:`, `feat(feature-name):`, `refactor:`). See `CONTRIBUTING.md`.
- [n/a] (UI changes only) Added screenshots/GIFs to description above
- [n/a] (UI changes only) Verified usability with NVDA/JAWS