chore: clean up unnecessary webdriverio/spectron references (#5342)
#### Details

This PR removes some unnecessary code and an unnecessary dependency on `webdriverio` that became obsolete when we switched the unified e2e tests from spectron to playwright. Per the comment in dependabot.yml, we were only using `webdriverio` for typings shared between it and `spectron`; it isn't required anymore.

I also removed a few other lingering bits of dead code related to Spectron while I was here (eg, a bunch of timeout constants that we stopped using with the move to playwright).

##### Motivation

Originally, this was part of a PR to resolve a CG alert against the `async` package. The non-dev dependencies on `async` were resolved by other PRs and the specific CG alert in question isn't applicable to dev dependencies, so it's no longer necessary to do any further resolutions for `async`; however, I kept this bit of the PR since removing the unnecessary code still seemed nice.

##### Context

n/a

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a" in the checkbox -->
- [n/a] Addresses an existing issue: #0000
- [x] Ran `yarn fastpass`
- [n/a] Added/updated relevant unit test(s) (and ran `yarn test`)
- [n/a] Verified code coverage for the changes made. Check coverage report at: `<rootDir>/test-results/unit/coverage`
- [x] PR title *AND* final merge commit title both start with a semantic tag (`fix:`, `chore:`, `feat(feature-name):`, `refactor:`). See `CONTRIBUTING.md`.
- [n/a] (UI changes only) Added screenshots/GIFs to description above
- [n/a] (UI changes only) Verified usability with NVDA/JAWS