feat(mv3-part-3): Move tabIdToContextMap out of TargetPageController (#5436)
#### Details

Moves tabIdToContextMap out of TargetPageController and into a new TabContextManager object

##### Motivation

This is part of a larger refactor of TargetPageController and ExtensionDetailsViewController, with the goal of simplifying the tangled dependency between them and removing duplicate tab event listeners that are registered in both.

##### Context

Currently, TargetPageController depends on ExtensionDetailsViewController for the sole purpose of passing ExtensionDetailsViewController a callback function, onDetailsViewTabRemoved. This function is only called from and used by ExtensionDetailsViewController, so it's strange for it to live in TargetPageController as a private method. It's there because it needs access to the TabContext to send a message the target page, and tabIdToContextMap lives in TargetPageController. Its only function is to use that TabContext's interpreter to interpret a message.

This PR moves tabIdToContextMap into another object, TabContextManager, which wraps the current uses of tabIdToContextMap. Currently, TargetPageController has sole access to and responsibility of TabContextManager, but a future PR will add it to ExtensionDetailsViewController so they don't need to depend on each other and pass callbacks back and forth.

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a" in the checkbox -->
- [ ] Addresses an existing issue: #0000
- [x] Ran `yarn fastpass`
- [x] Added/updated relevant unit test(s) (and ran `yarn test`)
- [x] Verified code coverage for the changes made. Check coverage report at: `<rootDir>/test-results/unit/coverage`
- [x] PR title *AND* final merge commit title both start with a semantic tag (`fix:`, `chore:`, `feat(feature-name):`, `refactor:`). See `CONTRIBUTING.md`.
- [ ] (UI changes only) Added screenshots/GIFs to description above
- [ ] (UI changes only) Verified usability with NVDA/JAWS

---

admin-merging to work around CLA bot issue