feat(mv3-part-3): DevToolsMonitor (#5465)
Creates DevToolsMonitor to poll active devtools for their status every 500ms until they close. It will only poll if there are active devtools. With this PR, DevToolsMonitor will poll alongside the browser connection that we currently use to detect devtools opening and closing, but the plan is for it to eventually replace that connection.