feat(mv3-part-4): Fix tab message errors (#5512)
#### Details

- In injected scripts, handle a scenario where the storeHub contains stores, but not all stores have their state initialized
- Change the signature of getAllStoreData() to return a Partial instead of assuming all storeData are present

##### Motivation

This fixes frequent error messages in the MV3 extension when a scan is started. The same error occurs silently in the MV2 extension, but does not affect the functionality of the extension.

##### Context

The cause of this bug is that the injected content scripts register some store update listeners that assume the store's state already exists, which isn't the case when the scripts are first initialized, so the initial store update message throws an error. This PR assumes that none of the listeners registered by ClientStoreListener need to be called before the store state is initialized (the listener that updates the StoreProxy state is called separately).