chore: turn on no-floating-promises for most directories (#5517)
Turns on the no-floating-promises eslint rule for all directories except background/ and injected/, and fixes the errors that were caught.

This PR fixes 26 issues caught by this rule out of 48 in total. The issues fixed in this PR are fairly minor, and the remaining issues in background/ and injected/ are more complex and involve changes that propagate more to other files. Those issues will be addressed in future PRs.