feat(mv3-part-4): Avoid using stale data in CardSelectionStore (#5549)
#### Details

- Make CardSelectionStoreData.rules nullable and set it to null by default
- Reset CardSelectionStoreData to the default state when the target tab URL changes
- Do not try to get a selectorMap if CardSelectionStoreData.rules is null

##### Motivation

This fixes an issue exposed by #5537. When the Automated checks or Needs review ad hoc visualizers are enabled, the DrawingController receives two requests, one with an empty list, and one with the correct list of elements to highlight. This is caused by two consecutive store updates, one to UnifiedScanResultStore and one to CardSelectionStore, which both trigger a visualization update. Between these two store updates, stale data in CardSelectionStore causes DrawingController to receive an empty list of elements. While working on #5537, we saw these two visualization requests racing in an E2E test, which caused the visualization to be wiped clean if the invalid request with the empty list was processed after the request with the correct element list.

This PR fixes the issue by setting CardSelectionStoreData.rules to null be default, to distinguish between situations where no scan has yet been run vs when a scan has completed and no issues were found. It also removes stale data from the store if the tab is reloaded or the URL changes. The visualization is not updated unless we have valid data from both the CardSelectionStore and the UnifiedScanResultStore.