fix(debug-tools): remove telemetry listener on viewer unmount (#5592)
#### Details

Today, when toggling the debug tools UI from the telemetry viewer back to the store viewer, it emits the following error:

```
Warning: Can't perform a React state update on an unmounted component. This is a no-op, but it indicates a memory leak in your application. To fix, cancel all subscriptions and asynchronous tasks in the componentWillUnmount method.
    in TelemetryViewer (created by CurrentView)
```

This happens because `TelemetryViewer.componentDidMount` registers a listener that `TelemetryViewer.componentWillUnmount` does not remove.

This PR fixes the warning by ensuring that the listener is removed on unmount of the viewer.

##### Motivation

Avoid unnecessary error spew

##### Context

n/a

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a" in the checkbox -->
- [n/a] Addresses an existing issue: #0000
- [x] Ran `yarn fastpass`
- [x] Added/updated relevant unit test(s) (and ran `yarn test`)
- [x] Verified code coverage for the changes made. Check coverage report at: `<rootDir>/test-results/unit/coverage`
- [x] PR title *AND* final merge commit title both start with a semantic tag (`fix:`, `chore:`, `feat(feature-name):`, `refactor:`). See `CONTRIBUTING.md`.
- [n/a] (UI changes only) Added screenshots/GIFs to description above
- [n/a] (UI changes only) Verified usability with NVDA/JAWS